basically working!

+++ monster_api/models.py
@@ -4,7 +4,7 @@ from django.db import models
 
 class Monster(models.Model):
     name = models.CharField(max_length=255)
-    height = models.IntegerField()
+    height = models.FloatField()
     description = models.TextField(max_length=1000)
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
